zip source now has version number

+++ src/_zip_source.py
@@ -2,7 +2,23 @@ import os
 import sys
 import platform
 
-zip_file_name = "duckypad_autoprofile_latest_source.zip"
+THIS_VERSION = None
+try:
+    mainfile = open('duckypad_autoprofile.py')
+    for line in mainfile:
+        if "THIS_VERSION_NUMBER =" in line:
+            THIS_VERSION = line.replace('\n', '').replace('\r', '').split("'")[-2]
+    mainfile.close()
+except Exception as e:
+    print('zip_source exception:', e)
+    exit()
+
+if THIS_VERSION is None:
+    print('could not find version number!')
+    exit()
+
+zip_file_name = f"duckypad_autoprofile_{THIS_VERSION}_source.zip"
+
 os.system("rm -fv ./*.zip")
 
 current_os = platform.system()
